refactor(auth): 重构认证模块为蓝图结构
将认证相关路由、模板和逻辑从主应用分离到auth蓝图
移动用户模型和数据库扩展到独立模块
更新静态资源引用路径为url_for格式
